% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_age_categorical.R
\name{add_AgeCategorical}
\alias{add_AgeCategorical}
\title{Add the age categorical in the data set}
\usage{
add_AgeCategorical(data)
}
\arguments{
\item{data}{data set}
}
\value{

}
\description{
Add the age categorical in the data set
}
