% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subj_analysis_set.R
\name{add_node_SbjAnaSet}
\alias{add_node_SbjAnaSet}
\title{Add the node of the subject per analysis set in the xml file which is in the category Baseline characteristics}
\usage{
add_node_SbjAnaSet(xml, data)
}
\arguments{
\item{xml}{takes as parameter a xml file}

\item{data}{takes as parameter a data frame with the ID of the patient,the center code, the age,
the sex, the randomisation group and the subject analysis group}
}
\value{
xml file
}
\description{
Add the node of the subject per analysis set in the xml file which is in the category Baseline characteristics
}
