% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{add_node_SubjectPerArm}
\alias{add_node_SubjectPerArm}
\title{Add the node of the subject per arm in the xml file which is in the category Subject disposition}
\usage{
add_node_SubjectPerArm(xml, data)
}
\arguments{
\item{xml}{takes as a parameter xml file}

\item{data}{takes as parameter a data frame with the ID of the patient,the center code,the age,
the sex, the randomisation group and the subject analysis group}
}
\value{
xml file
}
\description{
Add the node of the subject per arm in the xml file which is in the category Subject disposition
}
