% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbj_country_and_age.R
\name{add_node_SubPerCountry}
\alias{add_node_SubPerCountry}
\title{Add nodes "subjects" in "countSubjectCountry" in the category Trial information}
\usage{
add_node_SubPerCountry(xml, x)
}
\arguments{
\item{xml}{takes as parameter a xml file}

\item{x}{takes as parameter a function}
}
\value{
xml file
}
\description{
Add nodes "subjects" in "countSubjectCountry" in the category Trial information
}
