% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_age_categorical.R
\name{add_reportingGroup}
\alias{add_reportingGroup}
\title{Add the nodes of the reporting group in the XML file in the category Baseline characteristics}
\usage{
add_reportingGroup(xml, reportingGroupId)
}
\arguments{
\item{xml}{takes as parameter a XML file}

\item{reportingGroupId}{the id of the reporting group}
}
\value{

}
\description{
Add the nodes of the reporting group in the XML file in the category Baseline characteristics
}
