% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{addArm}
\alias{addArm}
\title{Add arms in the XML file which is in the category Subject disposition}
\usage{
addArm(xml, arm_id, title, description = "", arm_type)
}
\arguments{
\item{xml}{takes as parameter a XML file}

\item{arm_id}{takes as parameter an ID of the arm}

\item{title}{takes as parameter a name of the arm}

\item{description}{takes as parameter a description of the arm}

\item{arm_type}{takes as parameter a type of the arm}
}
\value{
XML file
}
\description{
Add arms in the XML file which is in the category Subject disposition
}
