% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{addArmBasedOnDictionnary}
\alias{addArmBasedOnDictionnary}
\title{Create the arms based on the dictionnary of the arms in the category Subject disposition}
\usage{
addArmBasedOnDictionnary(xml, dictionnary)
}
\arguments{
\item{xml}{takes as a parameter a XML file}

\item{dictionnary}{takes as parameter a dictionnary of the arm}
}
\value{

}
\description{
Create the arms based on the dictionnary of the arms in the category Subject disposition
}
