% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_age_categorical.R
\name{addIdCategory_AgeCategorical}
\alias{addIdCategory_AgeCategorical}
\title{Add the ID of the age categorical which is in the dictionnary of the category in the data set}
\usage{
addIdCategory_AgeCategorical(data, dictionnary)
}
\arguments{
\item{data}{the data set}

\item{dictionnary}{the dictionnary of the age categorical}
}
\value{

}
\description{
Add the ID of the age categorical which is in the dictionnary of the category in the data set
}
