% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_age_categorical.R
\name{addNodeReportModel}
\alias{addNodeReportModel}
\title{Add the nodes of the reporting model in the category Baseline characteristics}
\usage{
addNodeReportModel(xml, reportingModelType)
}
\arguments{
\item{xml}{takes as parameter a XML file}

\item{reportingModelType}{takes as parameter the type of baseline reporting model}
}
\value{
xml file
}
\description{
Add the nodes of the reporting model in the category Baseline characteristics
}
