% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{addPeriod}
\alias{addPeriod}
\title{Add the nodes of the Period in the XML file which is in the category Subject disposition}
\usage{
addPeriod(
  xml,
  mutuallyExclusiveArms,
  baselinePeriod,
  blinded,
  allocation_type,
  blinding_type
)
}
\arguments{
\item{xml}{takes as parameter a XML file}

\item{mutuallyExclusiveArms}{takes as parameter a boolean type}

\item{baselinePeriod}{takes as parameter a boolean type}

\item{blinded}{takes as parameter a boolean type}

\item{allocation_type}{takes as parameter the allocation type of the trial}

\item{blinding_type}{takes as parameter the blinding type of the trial}
}
\value{
XML file
}
\description{
Add the nodes of the Period in the XML file which is in the category Subject disposition
}
