% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subj_analysis_set.R
\name{addSbjAnaSetBasedOnDictionnary}
\alias{addSbjAnaSetBasedOnDictionnary}
\title{Create the subject analysis set based on the dictionnary of the subject analysis set}
\usage{
addSbjAnaSetBasedOnDictionnary(xml, dictionnary)
}
\arguments{
\item{xml}{takes as parameter a xml file}

\item{dictionnary}{takes as parameter a dictionnary of the subject analysis set}
}
\value{
xml file
}
\description{
Create the subject analysis set based on the dictionnary of the subject analysis set
}
