% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subj_analysis_set.R
\name{addsubjectAnalysisset}
\alias{addsubjectAnalysisset}
\title{Add the nodes of the subject analysis set which is in the category Baseline characteristics}
\usage{
addsubjectAnalysisset(xml, sbj_id, title, description = "", sbjAnalysisType)
}
\arguments{
\item{xml}{takes as parameter a XML file}

\item{title}{takes as parameter the name of the subject analysis set}

\item{description}{takes as parameter the description of the subject analysis set}

\item{sbjAnalysisType}{takes as parameter the type of the subject analysis set}

\item{id}{takes as parameter an ID of the subject analysis set}
}
\value{
xml file
}
\description{
Add the nodes of the subject analysis set which is in the category Baseline characteristics
}
