% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbj_country_and_age.R
\name{countAgeCat_per_years}
\alias{countAgeCat_per_years}
\title{Count the number of subjects per age range}
\usage{
countAgeCat_per_years(data)
}
\arguments{
\item{data}{takes as parameter a data frame with the ID of the patient,the center code,
age, sex, the randomisation group and the subject analysis group}
}
\value{
a data frame
}
\description{
Count the number of subjects per age range
}
