% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_age_categorical.R
\name{countAgeCategorical_per_arm}
\alias{countAgeCategorical_per_arm}
\title{Count the number of subject per arm and per age categorical in the cateogry Baseline characteristics}
\usage{
countAgeCategorical_per_arm(data)
}
\arguments{
\item{data}{takes as parameter the data set}
}
\value{

}
\description{
Count the number of subject per arm and per age categorical in the cateogry Baseline characteristics
}
