% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{countPerArm}
\alias{countPerArm}
\title{Count the number of the subject per arm}
\usage{
countPerArm(data)
}
\arguments{
\item{data}{takes as parameter a data frame with the ID of the patient,the center code,
age, sex, the randomisation group and the subject analysis group}
}
\value{
data frame
}
\description{
Count the number of the subject per arm
}
