% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subj_analysis_set.R
\name{countPerSubjAnaSet}
\alias{countPerSubjAnaSet}
\title{Count the number of subjects analysed per group}
\usage{
countPerSubjAnaSet(data)
}
\arguments{
\item{data}{takes as parameter a data frame with the ID of the patient, the center code, the age, the sex,
the randomisation group and the subject analysis group}
}
\value{
a data frame
}
\description{
Count the number of subjects analysed per group
}
