% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbj_country_and_age.R
\name{countSubjectCountry}
\alias{countSubjectCountry}
\title{Count the number of the subject per country}
\usage{
countSubjectCountry(data, dictionnary)
}
\arguments{
\item{data}{takes as parameter a data frame with the ID of the patient,the center code,
age, sex, the randomisation group and the subject analysis group}

\item{dictionnary}{takes as parameter a dictionnary of the country}
}
\value{
a data frame with the number of the subject per country
}
\description{
Count the number of the subject per country
}
