% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safety_summary.R
\name{create.safety_summary}
\alias{create.safety_summary}
\title{function that creates a safety_summary object from individual data.frames}
\usage{
create.safety_summary(group, non_serious, serious)
}
\arguments{
\item{group}{a data frame that contains the group-level statistics}

\item{non_serious}{a data frame that contains the non-serious term-group level statistics}

\item{serious}{a data frame that contains the serious term-group level statistics}
}
\value{
a safety_summary object
}
\description{
function that creates a safety_summary object from individual data.frames
}
