% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safety_summary.R
\name{df_to_char}
\alias{df_to_char}
\title{function to make a data frame be entirely character vectors}
\usage{
df_to_char(df)
}
\arguments{
\item{df}{a data frame}
}
\description{
function to make a data frame be entirely character vectors
}
\keyword{internal}
