% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subj_analysis_set.R
\name{dictionnary_sbAnaSet}
\alias{dictionnary_sbAnaSet}
\title{Convert the data frame into a dictionnary with the id of the subject analysis set}
\usage{
dictionnary_sbAnaSet(dictionnary)
}
\arguments{
\item{dictionnary}{takes as parameter a dictionnary of the subject analysis set}
}
\value{
data frame which is the new dictionnary
}
\description{
Convert the data frame into a dictionnary with the id of the subject analysis set
}
