% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{formatDictionnary}
\alias{formatDictionnary}
\title{Convert the data frame into a dictionnary with the id of the arm}
\usage{
formatDictionnary(dictionnary)
}
\arguments{
\item{dictionnary}{takes as parameter a dictionnary of the arm}
}
\value{
data frame which is the new dictionnary
}
\description{
Convert the data frame into a dictionnary with the id of the arm
}
