% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{fusiontable}
\alias{fusiontable}
\title{Fusionning two datas frame}
\usage{
fusiontable(dataA, dataB)
}
\arguments{
\item{dataA}{takes as a parameter a data frame with an key column "title"}

\item{dataB}{takes as a parameter a data frame with an key column "title"}
}
\value{
a data frame
}
\description{
Fusionning two datas frame
}
