% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_age_categorical.R
\name{getAgeCategory_baselineChar}
\alias{getAgeCategory_baselineChar}
\title{Get the age categorical in the xml file and put in the dictionnary which contain the id of the age categorical and the name of the age catogrical
in the category Baseline characteristics}
\usage{
getAgeCategory_baselineChar(xml)
}
\arguments{
\item{xml}{takes as parameter a XML file}
}
\value{
data frame which is the dictionnary
}
\description{
Get the age categorical in the xml file and put in the dictionnary which contain the id of the age categorical and the name of the age catogrical
in the category Baseline characteristics
}
