% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{getallocation_type}
\alias{getallocation_type}
\title{Get the allocation (type of the randomisation) of the trial}
\usage{
getallocation_type(alloc_type = NULL)
}
\arguments{
\item{alloc_type}{takes as parameter a type of randomization among the following list: 'Randomised-controlled',
'Non-randomised-controlled', 'Not applicable'}
}
\value{

}
\description{
Get the allocation (type of the randomisation) of the trial
}
