% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{getArm}
\alias{getArm}
\title{Get the structure of the arm}
\usage{
getArm(id, title, description = "", arm_type)
}
\arguments{
\item{id}{takes as parameter an ID of the arm}

\item{title}{takes as parameter a name of the arm}

\item{description}{takes as parameter a description of the arm}

\item{arm_type}{takes as parameter a type of the arm}
}
\value{

}
\description{
Get the structure of the arm
}
