% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{getarm_type}
\alias{getarm_type}
\title{Get the type of the arms}
\usage{
getarm_type(arm_type = NULL)
}
\arguments{
\item{arm_type}{takes as parameter a type of randomization among the following list: 'Experimental', 'Active comparator',
'Placebo', 'No IMP', 'Other'}
}
\value{

}
\description{
Get the type of the arms
}
