% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{getblinding_type}
\alias{getblinding_type}
\title{Get the blinding type of the trial}
\usage{
getblinding_type(blinding_type = NULL)
}
\arguments{
\item{blinding_type}{takes as parameter a type of bliniding among the following list: 'Double blind', 'Single blind', 'Not blinded'}
}
\value{

}
\description{
Get the blinding type of the trial
}
