% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_age_categorical.R
\name{getContainCountableValue}
\alias{getContainCountableValue}
\title{Get the nodes of the countable value of the age categorical in the XML file, in the category Baseline characteristics}
\usage{
getContainCountableValue(xml)
}
\arguments{
\item{xml}{takes as parameter a XML file}
}
\value{

}
\description{
Get the nodes of the countable value of the age categorical in the XML file, in the category Baseline characteristics
}
