% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{getReportingModelType}
\alias{getReportingModelType}
\title{Get the type of the baseline reporting model in the category Baseline characteristics}
\usage{
getReportingModelType(baselineReportingModel_type = NULL)
}
\arguments{
\item{baselineReportingModel_type}{takes as parameter a type of the baseline reporting model: "Arm" or "Period"}
}
\value{

}
\description{
Get the type of the baseline reporting model in the category Baseline characteristics
}
