% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{getsbjAnalysisset}
\alias{getsbjAnalysisset}
\title{Get the type of the subject analysis set}
\usage{
getsbjAnalysisset(sbj_type = NULL)
}
\arguments{
\item{sbj_type}{takes as parameter a type of randomization among the following list: 'Intention-to-treat', 'Per protocol', 'Full analysis', 'Safety analysis',
'Sub-group analysis', 'Modified intention-to-treat'}
}
\value{

}
\description{
Get the type of the subject analysis set
}
