% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbj_country_and_age.R
\name{recode_country}
\alias{recode_country}
\title{Recode the number of the centre with the code in the dictionnary of the country}
\usage{
recode_country(data, dictionnary)
}
\arguments{
\item{data}{takes as parameter a data frame with the ID of the patient,the center code,
the age, the sex, the randomisation group and the subject analysis group}

\item{dictionnary}{takes as parameter a dictionnary of the country}
}
\value{
data frame
}
\description{
Recode the number of the centre with the code in the dictionnary of the country
}
