% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbj_country_and_age.R
\name{replaceAgeRange}
\alias{replaceAgeRange}
\title{Replace the nodes of the age range with a new nodes which contain their values in the category Trial information}
\usage{
replaceAgeRange(xml, x)
}
\arguments{
\item{xml}{takes as parameter XML file}

\item{x}{takes as parameter a function which count the number of the subject per age range}
}
\value{
xml file
}
\description{
Replace the nodes of the age range with a new nodes which contain their values in the category Trial information
}
