% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_age_categorical.R
\name{replaceReadyforValue}
\alias{replaceReadyforValue}
\title{Replace the element of the node "Replace For Value" in in the category Baseline characteristics by true}
\usage{
replaceReadyforValue(xml)
}
\arguments{
\item{xml}{takes as parameter a XML file}
}
\value{

}
\description{
Replace the element of the node "Replace For Value" in in the category Baseline characteristics by true
}
