% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soc_code.R
\docType{data}
\name{soc_code}
\alias{soc_code}
\title{System Organ Class coding}
\format{a data frame with 3 columns and 27 rows
\describe{
   \item{soc_term}{a text description}
   \item{eutctId}{the eudraCT coding}
   \item{meddra}{the meddra code}
}}
\source{
\url{https://www.meddra.org/}, \url{https://spor.ema.europa.eu/rmswi/#/}
}
\usage{
soc_code
}
\description{
A dataset containing text descriptions and medDRA and EudraCT codes for each system organ class
}
\keyword{datasets}
